ignore zero prices and move impl types to own file
